Attempt to implement special requests in Frame2 (Stations selection screen)
Didn't really work (perhaps because deserialized objects != the actual objects?) Will try to implement in frame3 instead.

+++ src/Frame2.java
@@ -1,4 +1,5 @@
 import java.awt.EventQueue;
+import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
@@ -68,7 +69,7 @@ public class Frame2 extends JFrame {
 	public Frame2(MasterPlayerList playerList) {
 		setTitle("Stations");
 		stationDirectory = new File(stationDirectoryName);
-		stationList = new MasterStationList(stationDirectory);
+		stationList = new MasterStationList(stationDirectory, playerList);
 		originalDirectory = new File(this.stationOrigDirectoryName);
 		this.playerList = playerList;
 		initComponents();
@@ -80,6 +81,7 @@ public class Frame2 extends JFrame {
 			btnTemp.setForeground(Color.green);
 		} 
 		updateList();
+		stationList.updatePlayerLists();
 		
 	}
 	
@@ -132,8 +134,10 @@ public class Frame2 extends JFrame {
 		});
 		btnNewButton.addActionListener(new ActionListener() {//sets absents
 			public void actionPerformed(ActionEvent e) {
-				JlistStations.getSelectedValue().setDisabled(!JlistStations.getSelectedValue().isDisabled());
-				updateList();
+				if(JlistStations.getSelectedValue() != null) {
+					JlistStations.getSelectedValue().setDisabled(!JlistStations.getSelectedValue().isDisabled());
+					updateList();
+				}
 			}
 		});
 		btnBack.addActionListener(new ActionListener() {
@@ -143,8 +147,10 @@ public class Frame2 extends JFrame {
 		});
 		btnInfo.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				StationInfo info = new StationInfo(JlistStations.getSelectedValue());
-				info.setVisible(true);
+				if(JlistStations.getSelectedValue() != null) {
+					StationInfo info = new StationInfo(JlistStations.getSelectedValue());
+					info.setVisible(true);
+				}
 			}
 		});
 		btnDeleteAll.addActionListener(new ActionListener() {
@@ -157,6 +163,18 @@ public class Frame2 extends JFrame {
 				tempToggle();
 			}
 		});
+		
+		btnSpecRequest.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				if(JlistStations.getSelectedValue() != null) {
+					openSpecialRequest(JlistStations.getSelectedValue());
+				}
+			}
+		});
+	}
+	public void openSpecialRequest(Station selected) {
+		SpecialRequests request = new SpecialRequests(playerList, selected, stationList);
+		request.setVisible(true);
 	}
 	public void tempToggle() {
 		tempOn = !tempOn;
@@ -204,10 +222,12 @@ public class Frame2 extends JFrame {
 		}
 	}
 	public void deleteDir(File directory) {
-		for(File file : directory.listFiles()) {
-			file.delete();
+		if(directory.exists()) {
+			for(File file : directory.listFiles()) {
+				file.delete();
+			}
+			directory.delete();
 		}
-		directory.delete();
 	}
 	public void deleteAll() {
 		if((JOptionPane.showConfirmDialog(null, "ARE YOU ABSOLUTELY SURE YOU WANT TO DELETE ALL STATIONS?")) == JOptionPane.YES_OPTION); {
